requirements.txt
apispec==5.1.1
astroid==2.9.3
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
flake8==4.0.1
Flask==2.0.2
flask-marshmallow==0.14.0
flask-smorest==0.36.0
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
marshmallow==3.14.1
mccabe==0.6.1
packaging==21.3
platformdirs==2.4.1
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.12.2
pyparsing==3.0.7
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.8
webargs==8.1.0
Werkzeug==2.0.2
wrapt==1.13.3
